APIs

There are 4 functions that are exported, namely

Teukolsky_radial

GSN_radial

rstar_from_r

r_from_rstar

and there are 5 custom types that are exported, i.e.

BoundaryCondition

NormalizationConvention

Mode

GSNRadialFunction

TeukolskyRadialFunction

Currently, only the exported functions and types are documented below. Documentations for private (i.e. unexported) functions will be added at a later stage.

Functions

GeneralizedSasakiNakamura.Teukolsky_radial — Function

Teukolsky_radial(s::Int, l::Int, m::Int, a, omega, boundary_condition, rsin, rsout; horizon_expansion_order::Int=3, infinity_expansion_order::Int=6, data_type=Solutions._DEFAULTDATATYPE,  ODE_algorithm=Solutions._DEFAULTSOLVER, tolerance=Solutions._DEFAULTTOLERANCE)

Compute the Teukolsky function for a given mode (specified by s the spin weight, l the harmonic index, m the azimuthal index, a the Kerr spin parameter, and omega the frequency) and boundary condition (specified by boundary_condition which can be either IN for purely-ingoing at the horizon or UP for purely-outgoing at infinity) using the GSN formalism with GSN_radial.

The full GSN solution is converted to the corresponding Teukolsky solution $(R(r), dR/dr)$ and the incidence, reflection and transmission amplitude are converted from the GSN formalism to the Teukolsky formalism with the normalization convention that the transmission amplitude is normalized to 1 (i.e. normalization_convention=UNIT_TEUKOLSKY_TRANS).

GeneralizedSasakiNakamura.GSN_radial — Function

GSN_radial(s::Int, l::Int, m::Int, a, omega, boundary_condition, rsin, rsout; horizon_expansion_order::Int=3, infinity_expansion_order::Int=6, data_type=Solutions._DEFAULTDATATYPE,  ODE_algorithm=Solutions._DEFAULTSOLVER, tolerance=Solutions._DEFAULTTOLERANCE)

Compute the GSN function for a given mode (specified by s the spin weight, l the harmonic index, m the azimuthal index, a the Kerr spin parameter, and omega the frequency) and boundary condition (specified by boundary_condition which can be either IN for purely-ingoing at the horizon or UP for purely-outgoing at infinity).

The GSN function is numerically solved in the interval of tortoise coordinates $r_{*} \in$ [rsin, rsout] using the ODE solver (from DifferentialEquations.jl) specified by ODE_algorithm (default: Vern9()) with tolerance specified by tolerance (default: 1e-12). By default the data type used is ComplexF64 (i.e. double-precision floating-point number) but it can be changed by specifying data_type (e.g. Complex{BigFloat} for complex arbitrary precision number).

While the numerical GSN solution is only accurate in the range [rsin, rsout], the full GSN solution is constructed by smoothly attaching the asymptotic solutions near horizon (up to horizon_expansion_order-th order) and infinity (up to infinity_expansion_order-th order). Therefore, the now-semi-analytical GSN solution is accurate everywhere.

Return a GSNRadialFunction object which contains all the information about the GSN solution.

GeneralizedSasakiNakamura.Coordinates.rstar_from_r — Function

rstar_from_r(a, r)

Convert a Boyer-Lindquist coordinate r to the corresponding tortoise coordinate rstar.

GeneralizedSasakiNakamura.Coordinates.r_from_rstar — Function

r_from_rstar(a, rstar)

Convert a tortoise coordinate rstar to the corresponding Boyer-Lindquist coordiante r. It uses a bisection method when rstar <= 0, and Newton method otherwise.

The function assumes that $r \geq r_{+}$ where $r_{+}$ is the outer event horizon.

Types

BoundaryCondition

This is an enum type that can take either one of the two values

value

IN	purely ingoing at the horizon

UP	purely outgoing at infinity

NormalizationConvention

This is an enum type that can take either one of the two values

value

UNIT_GSN_TRANS	normalized to have a unit transmission amplitude for the GSN function

UNIT_TEUKOLSKY_TRANS	normalized to have a unit transmission amplitude for the Teukolsky function

Mode

This is a composite struct type that stores information about a mode

field

s	spin weight $s$

l	harmonic index $\ell$

m	azimuthal index $m$

a	Kerr spin parameter $a/M$

omega	frequency $M\omega$

lambda	spin-weighted spheroidal eigenvalue $\lambda$

GSNRadialFunction

This is a composite struct type that stores the output from GSN_radial

Tip

GSNRadialFunction(rstar) is equivalent to GSNRadialFunction.GSN_solution(rstar)[1], returning only the value of the GSN function evaluated at the tortoise coordinate rstar

field

mode	a Mode object storing information about the mode

boundary_condition	a BoundaryCondition object storing which boundary condition this function satisfies

rsin	numerical inner boundary $r_{*}^{\mathrm{in}}/M$ where the GSN equation is numerically evolved ($r_{*}$ is a tortoise coordinate)

rsout	numerical outer boundary $r_{*}^{\mathrm{out}}/M$ where the GSN equation is numerically evolved ($r_{*}$ is a tortoise coordinate)

horizon_expansion_order	order of the asymptotic expansion at the horizon

infinity_expansion_order	order of the asymptotic expansion at infinity

transmission_amplitude	transmission amplitude in the GSN formalism of this function

incidence_amplitude	incidence amplitude in the GSN formalism of this function

reflection_amplitude	reflection amplitude in the GSN formalism of this function

numerical_GSN_solution	numerical solution (ODESolution object from DifferentialEquations.jl) to the GSN equation in [rsin, rsout] if applicable; output is a vector $[ \hat{X}(r_{*}), d\hat{X}(r_{*})/dr_{*} ]$

numerical_Riccati_solution	numerical solution (ODESolution object from DifferentialEquations.jl) to the GSN equation in the Riccati form if applicable; output is a vector $[ \hat{\Phi}(r_{*}), d\hat{\Phi}(r_{*})/dr_{*} ]$

GSN_solution	full GSN solution where asymptotic solutions are smoothly attached; output is a vector $[ \hat{X}(r_{*}), d\hat{X}(r_{*})/dr_{*} ]$

normalization_convention	a NormalizationConvention object storing which normalization convention this function adheres to

TeukolskyRadialFunction

This is a composite struct type that stores the output from Teukolsky_radial

Tip

TeukolskyRadialFunction(r) is equivalent to TeukolskyRadialFunction.Teukolsky_solution(r)[1], returning only the value of the Teukolsky function evaluated at the Boyer-Lindquist coordinate r

field

mode	a Mode object storing information about the mode

boundary_condition	a BoundaryCondition object storing which boundary condition this function satisfies

transmission_amplitude	transmission amplitude in the Teukolsky formalism of this function

incidence_amplitude	incidence amplitude in the Teukolsky formalism of this function

reflection_amplitude	reflection amplitude in the Teukolsky formalism of this function

GSN_solution	a GSNRadialFunction object storing the corresponding GSN function

Teukolsky_solution	Teukolsky solution where asymptotic solutions are smoothly attached; output is a vector $[ \hat{R}(r), d\hat{R}(r)/dr ]$

normalization_convention	a NormalizationConvention object storing which normalization convention this function adheres to